xp and multi to double instead long
changed the db, so its accepts REAL, not BIGINT.

+++ src/main/java/dk/martinersej/generator/generator/GeneratorType.java
@@ -636,9 +636,9 @@ public enum GeneratorType {
         ;
 
         private final double price;
-        private final long xp;
+        private final double xp;
 
-        DropPrice(double price, long xp) {
+        DropPrice(double price, double xp) {
             this.price = price;
             this.xp = xp;
         }
@@ -647,7 +647,7 @@ public enum GeneratorType {
             return price;
         }
 
-        public long getXp() {
+        public double getXp() {
             return xp;
         }
     }
